[RW-4816][Risk=no] Add unordered list outside component
