+++ ui/src/app/components/lists.tsx
@@ -1,0 +1,15 @@
+import * as React from 'react';
+
+export const CustomBulletList = ({children}) => {
+  return <ul style={{position: 'relative', listStyle: 'none', marginLeft: 0}}>{children}</ul>;
+};
+export const CustomBulletListItem = ({bullet, children}) => {
+  return <React.Fragment>
+    <div style={{position: 'absolute', left: '-.2em'}}> {bullet} </div>
+    <li style={{marginLeft: '1.2em'}}> {children} </li>
+  </React.Fragment>;
+};
+
+export const OutsideUnorderedList = ({style = {}, children}) => {
+  return <ul style={{listStylePosition: 'outside', marginLeft: '1rem', ...style}}>{children}</ul>
+}
